+++ sdk/formrecognizer/azure-ai-formrecognizer/azure/ai/formrecognizer/_response_handlers.py
@@ -8,14 +8,14 @@
 
 from ._models import (
     USReceipt,
-    USReceiptType,
+    ReceiptType,
     FormField,
     USReceiptItem,
     FormPage,
     FormLine,
     FormTable,
     FormTableCell,
-    PageRange,
+    FormPageRange,
     RecognizedForm
 )
 
@@ -29,7 +29,7 @@ def prepare_us_receipt(response):
     for page in document_result:
         if page.fields is None:
             receipt = USReceipt(
-                page_range=PageRange(first_page=page.page_range[0], last_page=page.page_range[1]),
+                page_range=FormPageRange(first_page=page.page_range[0], last_page=page.page_range[1]),
                 pages=form_page[page.page_range[0]-1:page.page_range[1]],
                 form_type=page.doc_type,
             )
@@ -47,7 +47,7 @@ def prepare_us_receipt(response):
                 page.fields.get("MerchantPhoneNumber"),
                 read_result,
             ),
-            receipt_type=USReceiptType._from_generated(page.fields.get("ReceiptType")),
+            receipt_type=ReceiptType._from_generated(page.fields.get("ReceiptType")),
             receipt_items=USReceiptItem._from_generated(
                 page.fields.get("Items"), read_result
             ),
@@ -65,7 +65,7 @@ def prepare_us_receipt(response):
             transaction_time=FormField._from_generated(
                 "TransactionTime", page.fields.get("TransactionTime"), read_result
             ),
-            page_range=PageRange(
+            page_range=FormPageRange(
                 first_page=page.page_range[0], last_page=page.page_range[1]
             ),
             pages=form_page[page.page_range[0]-1:page.page_range[1]],
@@ -132,7 +132,7 @@ def prepare_unlabeled_result(response):
         if unlabeled_fields:
             unlabeled_fields = {field.name: field for field in unlabeled_fields}
         form = RecognizedForm(
-            page_range=PageRange(
+            page_range=FormPageRange(
                 first_page=page.page,
                 last_page=page.page
             ),
@@ -152,7 +152,7 @@ def prepare_labeled_result(response, model_id):
     result = []
     for doc in response.analyze_result.document_results:
         form = RecognizedForm(
-            page_range=PageRange(
+            page_range=FormPageRange(
                 first_page=doc.page_range[0],
                 last_page=doc.page_range[1]
             ),