+++ sdk/formrecognizer/azure-ai-formrecognizer/samples/async_samples/sample_train_model_with_labels_async.py
@@ -36,17 +36,16 @@ class TrainModelWithLabelsSampleAsync(object):
     container_sas_url = os.environ["CONTAINER_SAS_URL"]
 
     async def train_model_with_labels(self):
-        # [START create_form_training_client_async]
         from azure.ai.formrecognizer.aio import FormTrainingClient
         from azure.core.credentials import AzureKeyCredential
 
         form_training_client = FormTrainingClient(
             endpoint=self.endpoint, credential=AzureKeyCredential(self.key)
         )
-        # [END create_form_training_client_async]
-        async with form_training_client:
 
+        async with form_training_client:
             model = await form_training_client.train_model(self.container_sas_url, use_training_labels=True)
+
             # Custom model information
             print("Model ID: {}".format(model.model_id))
             print("Status: {}".format(model.status))