+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/test_custom_forms_from_url.py
@@ -23,20 +23,20 @@ class TestCustomFormsFromUrl(FormRecognizerTest):
     def test_custom_form_url_bad_endpoint(self, resource_group, location, form_recognizer_account, form_recognizer_account_key):
         with self.assertRaises(ServiceRequestError):
             client = FormRecognizerClient("http://notreal.azure.com", AzureKeyCredential(form_recognizer_account_key))
-            result = client.begin_recognize_custom_forms_from_url(model_id="xx", url=self.form_url_jpg)
+            result = client.begin_recognize_custom_forms_from_url(model_id="xx", form_url=self.form_url_jpg)
 
     @GlobalFormRecognizerAccountPreparer()
     def test_url_authentication_bad_key(self, resource_group, location, form_recognizer_account, form_recognizer_account_key):
         client = FormRecognizerClient(form_recognizer_account, AzureKeyCredential("xxxx"))
         with self.assertRaises(ClientAuthenticationError):
-            result = client.begin_recognize_custom_forms_from_url(model_id="xx", url=self.form_url_jpg)
+            result = client.begin_recognize_custom_forms_from_url(model_id="xx", form_url=self.form_url_jpg)
 
     @GlobalFormRecognizerAccountPreparer()
     def test_passing_bad_url(self, resource_group, location, form_recognizer_account, form_recognizer_account_key):
         client = FormRecognizerClient(form_recognizer_account, AzureKeyCredential(form_recognizer_account_key))
 
         with self.assertRaises(HttpResponseError):
-            poller = client.begin_recognize_custom_forms_from_url(model_id="xx", url="https://badurl.jpg")
+            poller = client.begin_recognize_custom_forms_from_url(model_id="xx", form_url="https://badurl.jpg")
 
     @GlobalFormRecognizerAccountPreparer()
     def test_pass_stream_into_url(self, resource_group, location, form_recognizer_account, form_recognizer_account_key):
@@ -46,7 +46,7 @@ class TestCustomFormsFromUrl(FormRecognizerTest):
             with self.assertRaises(HttpResponseError):
                 poller = client.begin_recognize_custom_forms_from_url(
                     model_id="xxx",
-                    url=fd,
+                    form_url=fd,
                 )
 
     @GlobalFormRecognizerAccountPreparer()
@@ -60,7 +60,7 @@ class TestCustomFormsFromUrl(FormRecognizerTest):
         with self.assertRaises(HttpResponseError):
             poller = fr_client.begin_recognize_custom_forms_from_url(
                 model.model_id,
-                url="https://badurl.jpg"
+                form_url="https://badurl.jpg"
             )
             form = poller.result()
 
@@ -80,7 +80,6 @@ class TestCustomFormsFromUrl(FormRecognizerTest):
         for label, field in form[0].fields.items():
             self.assertIsNotNone(field.confidence)
             self.assertIsNotNone(field.name)
-            self.assertIsNotNone(field.page_number)
             self.assertIsNotNone(field.value)
             self.assertIsNotNone(field.value_data.text)
             self.assertIsNotNone(field.label_data.text)
@@ -105,7 +104,6 @@ class TestCustomFormsFromUrl(FormRecognizerTest):
             for label, field in form.fields.items():
                 self.assertIsNotNone(field.confidence)
                 self.assertIsNotNone(field.name)
-                self.assertIsNotNone(field.page_number)
                 self.assertIsNotNone(field.value)
                 self.assertIsNotNone(field.value_data.text)
                 self.assertIsNotNone(field.label_data.text)
@@ -126,7 +124,6 @@ class TestCustomFormsFromUrl(FormRecognizerTest):
         for label, field in form[0].fields.items():
             self.assertIsNotNone(field.confidence)
             self.assertIsNotNone(field.name)
-            self.assertIsNotNone(field.page_number)
             self.assertIsNotNone(field.value_data.text)
             self.assertIsNotNone(field.value_data.bounding_box)
 
@@ -153,7 +150,6 @@ class TestCustomFormsFromUrl(FormRecognizerTest):
             for label, field in form.fields.items():
                 self.assertIsNotNone(field.confidence)
                 self.assertIsNotNone(field.name)
-                self.assertIsNotNone(field.page_number)
                 self.assertIsNotNone(field.value_data.text)
                 self.assertIsNotNone(field.value_data.bounding_box)