+++ sdk/search/azure-search-documents/azure/search/documents/_service/_models.py
@@ -3,10 +3,10 @@
 # Licensed under the MIT License. See License.txt in the project root for
 # license information.
 # --------------------------------------------------------------------------
-from ._generated.models import Analyzer, Tokenizer
+from ._generated.models import LexicalAnalyzer, LexicalTokenizer
 
 
-class PatternAnalyzer(Analyzer):
+class PatternAnalyzer(LexicalAnalyzer):
     """Flexibly separates text into terms via a regular expression.
     This analyzer is implemented using Apache Lucene.
 
@@ -49,7 +49,7 @@ class PatternAnalyzer(Analyzer):
         self.stopwords = kwargs.get("stopwords", None)
 
 
-class PatternTokenizer(Tokenizer):
+class PatternTokenizer(LexicalTokenizer):
     """Tokenizer that uses regex pattern matching to construct distinct tokens.
     This tokenizer is implemented using Apache Lucene.