+++ sdk/search/azure-search-documents/azure/search/documents/_service/aio/_synonym_maps_client.py
@@ -16,7 +16,6 @@ from ..._version import SDK_MONIKER
 
 if TYPE_CHECKING:
     # pylint:disable=unused-import,ungrouped-imports
-    from .._generated.models import Skill
     from typing import Any, Dict, List, Sequence, Union, Optional
     from azure.core.credentials import AzureKeyCredential
 
@@ -131,18 +130,15 @@ class SearchSynonymMapsClient(HeadersMixin):
         """
         kwargs["headers"] = self._merge_client_headers(kwargs.get("headers"))
         error_map, access_condition = get_access_conditions(
-            synonym_map,
-            kwargs.pop('match_condition', MatchConditions.Unconditionally)
+            synonym_map, kwargs.pop("match_condition", MatchConditions.Unconditionally)
         )
+        kwargs.update(access_condition)
         try:
             name = synonym_map.name
         except AttributeError:
             name = synonym_map
         await self._client.synonym_maps.delete(
-            synonym_map_name=name,
-            access_condition=access_condition,
-            error_map=error_map,
-            **kwargs
+            synonym_map_name=name, error_map=error_map, **kwargs
         )
 
     @distributed_trace_async
@@ -191,9 +187,9 @@ class SearchSynonymMapsClient(HeadersMixin):
         """
         kwargs["headers"] = self._merge_client_headers(kwargs.get("headers"))
         error_map, access_condition = get_access_conditions(
-            synonym_map,
-            kwargs.pop('match_condition', MatchConditions.Unconditionally)
+            synonym_map, kwargs.pop("match_condition", MatchConditions.Unconditionally)
         )
+        kwargs.update(access_condition)
         try:
             name = synonym_map.name
             if synonyms:
@@ -205,7 +201,6 @@ class SearchSynonymMapsClient(HeadersMixin):
         result = await self._client.synonym_maps.create_or_update(
             synonym_map_name=name,
             synonym_map=synonym_map,
-            access_condition=access_condition,
             error_map=error_map,
             **kwargs
         )