+++ sdk/core/azure-core/CHANGELOG.md
@@ -3,6 +3,13 @@
 
 ## 1.5.1 (Unreleased)
 
+### Bug fixes
+
+- Fix AttributeException in StreamDownloadGenerator #11462
+
+### Features
+
+- Added support for changesets as part of multipart message support #10485
 
 ## 1.5.0 (2020-05-04)